requirements.txt
pillow
exif
geopy
shapely
opencv-python
numpy
pathlib